# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "U1R API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payments": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create Payment",
                "operationId": "5e907b772c6a27dc7dafc6c76a5ffa16",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {},
                                    "transection_no": {},
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "status": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment created Successfully"
                    }
                }
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get Payment By ID",
                "operationId": "30f448f98368b8382bd347740dbb9fab",
                "responses": {
                    "200": {
                        "description": "Get Payment successfully"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Payment"
                ],
                "summary": "Update Payment",
                "operationId": "3f8b0efd784602e0a1d60a628ca0c831",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment Updated Successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}